refactor: isolate API router and mount v1 alias
